无敌了，三天完成1.21.1到1.21.5的移植

+++ src/main/java/com/mafuyu33/neomafishmod/mixin/enchantmentitemmixin/luoyangshovel/FallingBlockEntityMixin.java
@@ -36,7 +36,7 @@ public abstract class FallingBlockEntityMixin extends Entity {
 				if (boundingBox.intersects(livingEntity.getBoundingBox())) {
 					// 对生物实体造成伤害
 					DamageSource damageSource = damageSources().fallingBlock(fallingBlockEntity);
-					livingEntity.hurt(damageSource, fallingBlockEntity.fallDistance);
+					livingEntity.hurt(damageSource, (float) fallingBlockEntity.fallDistance);
 				}
 			}
 		}
